Implemented a animation controller for the hud avatar. Fixed an issue with the 3D sound (dumb int cast).
git-svn-id: https://dev.wibe.media/svn/code/java/trunk/litiengine@1365 16493bb7-3577-b145-b9c3-72a97d279fb6

+++ src/de/gurkenlabs/litiengine/sound/PaulsSoundEngine.java
@@ -70,8 +70,8 @@ public class PaulsSoundEngine extends SoundEngine {
 
     // center sounds that are roughly played at the same location
 
-    float locationX = (int) this.getListenerPosition().getX();
-    float locationY = (int) this.getListenerPosition().getY();
+    float locationX = (float) this.getListenerPosition().getX();
+    float locationY = (float) this.getListenerPosition().getY();
     if (this.getMaxListenerRadius() != 0 && entity.getDimensionCenter().distance(this.getListenerPosition()) > this.getMaxListenerRadius()) {
       locationX = 10000;
       locationY = 10000;
